refactor: move tests into falcon package dir
