+++ falcon/benchmarks/README.md
@@ -1,0 +1,3 @@
+# Benchmarks
+
+To optimize and to track degradations over time.
