+++ falcon/package.json
@@ -7,7 +7,9 @@
   "license": "MIT",
   "scripts": {
     "build": "tsc",
-    "format": "prettier --write src"
+    "format": "prettier --write src",
+    "jest": "NODE_OPTIONS=--experimental-vm-modules npx jest",
+    "test": "yarn jest"
   },
   "dependencies": {
     "@duckdb/duckdb-wasm": "^1.20.0",
@@ -29,6 +31,8 @@
     "@types/ndarray": "1.0.11",
     "@types/ndarray-ops": "^1.2.4",
     "prettier": "^2.8.1",
-    "typescript": "^4.9.4"
+    "typescript": "^4.9.4",
+    "jest": "^29.3.1",
+    "ts-jest": "^29.0.3"
   }
 }
