+++ falcon/tsconfig.json
@@ -1,25 +1,25 @@
 {
-    "compilerOptions": {
-        "module": "esnext",
-        "target": "es2021",
-        "lib": ["es2021", "dom"],
-        "outDir": "./build",
-        "rootDir": "./src",
-        "moduleResolution": "node",
-        "declaration": true,
-        "esModuleInterop": true,
-        "noImplicitAny": false,
-        "strictNullChecks": true,
-        "noImplicitReturns": true,
-        "noImplicitThis": true,
-        "noUnusedLocals": true,
-        "noUnusedParameters": true,
-        "sourceMap": true,
-        "allowSyntheticDefaultImports": true,
-        "skipLibCheck": true,
-        "resolveJsonModule": true,
-        "isolatedModules": true,
-        "noEmit": false
-    },
-    "include": ["src"]
+  "compilerOptions": {
+    "module": "esnext",
+    "target": "es2021",
+    "lib": ["es2021", "dom"],
+    "outDir": "./build",
+    "rootDir": "./src",
+    "moduleResolution": "node",
+    "declaration": true,
+    "esModuleInterop": true,
+    "noImplicitAny": false,
+    "strictNullChecks": true,
+    "noImplicitReturns": true,
+    "noImplicitThis": true,
+    "noUnusedLocals": true,
+    "noUnusedParameters": true,
+    "sourceMap": true,
+    "allowSyntheticDefaultImports": true,
+    "skipLibCheck": true,
+    "resolveJsonModule": true,
+    "isolatedModules": true,
+    "noEmit": false
+  },
+  "include": ["src"]
 }
