+++ package.json
@@ -14,16 +14,11 @@
     "start:flights-arrow": "yarn workspace flights-arrow dev --open",
     "start:flights-duckdb": "yarn workspace flights-duckdb dev --open",
     "dev": "yarn start:flights-arrow",
-    "jest": "NODE_OPTIONS=--experimental-vm-modules npx jest",
-    "test": "yarn jest",
-    "build": "yarn workspace falcon build",
-    "format": "yarn workspace falcon format"
+    "build": "yarn workspace falcon-vis build",
+    "format": "yarn workspace falcon-vis format",
+    "test": "yarn workspace falcon-vis test"
   },
-  "devDependencies": {
-    "jest": "^29.3.1",
-    "ts-jest": "^29.0.3",
-    "typescript": "^4.9.4"
-  },
+  "devDependencies": {},
   "directories": {
     "examples": "./examples/*",
     "falcon": "./falcon"
